+++ src/endpoints/accounts/entities/account.detailed.ts
@@ -1,5 +1,5 @@
 import { ComplexityEstimation } from "@multiversx/sdk-nestjs-common";
-import { Field, ObjectType } from "@nestjs/graphql";
+import { Field, Float, ObjectType } from "@nestjs/graphql";
 import { ApiProperty } from "@nestjs/swagger";
 import { ScamInfo } from "src/common/entities/scam-info.dto";
 import { NftCollectionAccount } from "src/endpoints/collections/entities/nft.collection.account";
@@ -36,6 +36,14 @@ export class AccountDetailed extends Account {
   @Field(() => String, { description: 'Owner address for the given detailed account.' })
   @ApiProperty({ description: 'The address in bech 32 format of owner account' })
   ownerAddress: string = '';
+
+  @Field(() => Float, { description: 'Transactions count for the given detailed account.' })
+  @ApiProperty({ description: 'The number of transactions performed on this account' })
+  txCount?: number;
+
+  @Field(() => Float, { description: 'Smart contract results count for the given detailed account.' })
+  @ApiProperty({ description: 'The number of smart contract results of this account' })
+  scrCount: number = 0;
 
   @Field(() => Boolean, { description: 'If the given detailed account is upgradeable.', nullable: true })
   @ApiProperty({ description: 'Specific property flag for smart contract', type: Boolean })
