+++ src/endpoints/accounts/entities/account.query.options.ts
@@ -1,5 +1,6 @@
 import { SortOrder } from "src/common/entities/sort.order";
 import { AccountSort } from "./account.sort";
+import { BadRequestException } from "@nestjs/common";
 
 export class AccountFilter {
   constructor(init?: Partial<AccountFilter>) {
@@ -13,6 +14,12 @@ export class AccountFilter {
   isSmartContract?: boolean;
   withOwnerAssets?: boolean;
   withDetails?: boolean;
+
+  validate(size: number) {
+    if (this.withDetails && size > 25) {
+      throw new BadRequestException('Size must be less than 25 when withDetails is set');
+    }
+  }
 
   isSet(): boolean {
     return this.ownerAddress !== undefined ||
